[Rename] modules/aggs-matrix-stats
This PR refactors the aggs-matrix-stats module as part of the Elasticsearch to OpenSearch renaming.
