--- modules/aggs-matrix-stats/src/main/resources/META-INF/services/org.elasticsearch.plugins.spi.NamedXContentProvider
@@ -1,1 +1,0 @@
-org.elasticsearch.search.aggregations.matrix.spi.MatrixStatsNamedXContentProvider
