+++ modules/aggs-matrix-stats/src/main/resources/META-INF/services/org.opensearch.plugins.spi.NamedXContentProvider
@@ -1,0 +1,1 @@
+org.opensearch.search.aggregations.matrix.spi.MatrixStatsNamedXContentProvider
